refactor: style 파일명 styles로 통일, useAxios 파일명 수정
