+++ src/components/Divider/Divider.styles.ts
